requirements.txt
# =====================================
# CENTRALLY - Requirements
# Plataforma de Analítica de Transporte Urbano
# Alcaldía de Medellín
# =====================================

# ----------------------------------
# Core Python Data Stack
# ----------------------------------
pandas==2.1.4
numpy==1.26.2
scipy==1.11.4

# ----------------------------------
# Capa 1: Ingesta de Datos
# ----------------------------------
requests==2.31.0
beautifulsoup4==4.12.2
openpyxl==3.1.2          # Excel files
pandera==0.17.2          # Data validation
pydantic==2.5.3          # Schema validation

# ----------------------------------
# Capa 2: Transformación ETL/ELT
# ----------------------------------
dbt-core==1.7.4
dbt-postgres==1.7.4
dbt-duckdb==1.7.1
sqlalchemy==2.0.25
alembic==1.13.1          # DB migrations

# ----------------------------------
# Capa 3: Almacenamiento
# ----------------------------------
psycopg2-binary==2.9.9   # PostgreSQL adapter
duckdb==0.9.2
geoalchemy2==0.14.3      # PostGIS support

# ----------------------------------
# Geospatial Data
# ----------------------------------
geopandas==0.14.1
shapely==2.0.2
fiona==1.9.5
pyproj==3.6.1
folium==0.15.1

# ----------------------------------
# Capa 4: Analítica y ML
# ----------------------------------
scikit-learn==1.3.2
xgboost==2.0.3
lightgbm==4.1.0
statsmodels==0.14.1
prophet==1.1.5           # Time series forecasting
mlflow==2.9.2            # ML lifecycle management

# ----------------------------------
# Capa 5: LLMs y RAG
# ----------------------------------
openai==1.6.1
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.2
faiss-cpu==1.7.4         # Vector store
chromadb==0.4.22
sentence-transformers==2.2.2
tiktoken==0.5.2          # Token counting

# ----------------------------------
# Capa 6: Visualización (Dash)
# ----------------------------------
dash==2.14.2
dash-bootstrap-components==1.5.0
dash-leaflet==1.0.15     # Interactive maps
plotly==5.18.0
kaleido==0.2.1           # Static image export

# ----------------------------------
# Capa 7: Backend (FastAPI)
# ----------------------------------
fastapi==0.108.0
uvicorn[standard]==0.25.0
gunicorn==21.2.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0  # JWT
passlib[bcrypt]==1.7.4   # Password hashing
python-multipart==0.0.6  # Form data

# ----------------------------------
# Capa 8: Gobernanza de Datos
# ----------------------------------
great-expectations==0.18.8
apache-airflow==2.8.0    # Orchestration (optional)
prefect==2.14.11         # Alternative orchestration

# ----------------------------------
# Development & Testing
# ----------------------------------
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
black==23.12.1           # Code formatting
flake8==7.0.0            # Linting
isort==5.13.2            # Import sorting
mypy==1.7.1              # Type checking
pre-commit==3.6.0

# ----------------------------------
# Utilities
# ----------------------------------
python-dotenv==1.0.0     # Environment variables
loguru==0.7.2            # Advanced logging
tqdm==4.66.1             # Progress bars
click==8.1.7             # CLI
pyyaml==6.0.1
jinja2==3.1.2
pytz==2023.3.post1
python-dateutil==2.8.2

# ----------------------------------
# Monitoring & Observability
# ----------------------------------
prometheus-client==0.19.0
sentry-sdk==1.39.1

# ----------------------------------
# Data Quality
# ----------------------------------
cerberus==1.3.5
jsonschema==4.20.0
